+++ tests/test_describe_command.rs
@@ -64,4 +64,17 @@ fn test_describe() {
         .assert()
         .failure();
     assert!(get_stderr_string(&assert).contains("Failed to run"));
+
+    // `ui.editor` config overrides `$EDITOR`
+    std::fs::write(&edit_script, "").unwrap();
+    test_env.add_config(
+        br#"[ui]
+    editor = "bad-editor-from-config""#,
+    );
+    let assert = test_env
+        .jj_cmd(&repo_path, &["describe"])
+        .env("EDITOR", "bad-editor-from-env")
+        .assert()
+        .failure();
+    assert!(get_stderr_string(&assert).contains("bad-editor-from-config"));
 }